refactor(app): extract chat projection
